Added a client_provider module
This is for restructuring the code a little.  We basically create a
simple class which listens to the AvahiBrowser and maintains a list of
discovered clients.  This is very generic.  So you can get an accurate
list of providers of any service.  Hence, this should probably go to the
network module at some stage.

However, there is a second, more specific class for the use with GNOME
Keysign.  It returns the list of discovered services in a way which is
expected by the application.